Flytte sjekk av grunnlagsversjon til egen hjelpeklasse
